Defined command line options.
