+++ README.md
@@ -1,19 +1,19 @@
 PLEN - Control Server
 ================================================================================
-Copyright (c) 2015.
+Copyright (c) 2015,
 ---
 - [Kazuyuki TAKASE](https://github.com/Guvalif)
 - [PLEN Project Company Ltd.](http://plen.jp)
 
-Build enviroment.
+Build Enviroment
 ---
-- Windows 8.1 Professional edition.
-- Windows 7 Home Premium 64bit SP1
+- Windows 8.1 Professional 64bit
+- Windows 7 Home Premium SP1 64bit
 - Python 2.7.9
 - Flask 0.10.1
 - pySerial 2.7
 - py2exe 0.6.9
 
-License.
+License
 ---
 This software is released under the MIT License.
