+++ control_server/device_map.json
@@ -1,0 +1,20 @@
+{
+	"left_elbow_roll":      3,
+	"left_foot_pitch":      7,
+	"left_foot_roll":       8,
+	"left_knee_pitch":      6,
+	"left_shoulder_pitch":  0,
+	"left_shoulder_roll":   2,
+	"left_thigh_pitch":     5,
+	"left_thigh_roll":      4,
+	"left_thigh_yaw":       1,
+	"right_elbow_roll":     15,
+	"right_foot_pitch":     19,
+	"right_foot_roll":      20,
+	"right_knee_pitch":     18,
+	"right_shoulder_pitch": 12,
+	"right_shoulder_roll":  14,
+	"right_thigh_pitch":    17,
+	"right_thigh_roll":     16,
+	"right_thigh_yaw":      13
+}
