+++ control_server/scripts/setup.sh
